Move test models into `tests/data`
Ref. #252
