+++ tests/unit/parser/parser_test.py
@@ -10,8 +10,8 @@ from rflx import expression as expr, model
 from rflx.error import Location, RecordFluxError, Severity, Subsystem, fail
 from rflx.identifier import ID
 from rflx.parser import ast, cache, parser
-from tests import models
 from tests.const import EX_SPEC_DIR, SPEC_DIR
+from tests.data import models
 
 
 def assert_specifications_files(
